[DEV-8437] Initial changes.
